feat(commands): enhance session command with prompt_toolkit goodies
removed duplicated session commands
added a prompt_toolkit session to get autocompletion, autosuggestion, command history and more
