+++ tests/commands/session/test_pong.py
@@ -1,7 +1,7 @@
 import pytest
 from trio_websocket import serve_websocket
 
-from tests.helpers import get_fake_input, server_handler
+from tests.helpers import server_handler
 from ws.commands.session import main
 
 
@@ -10,9 +10,9 @@ from ws.commands.session import main
     [('pong foo bar', 'Unknown argument: bar\n'), ('pong foo bar tar', 'Unknown arguments: bar tar\n')],
 )
 async def test_should_print_unknown_arguments_when_they_are_passed_to_pong_sub_command(
-    capsys, mocker, nursery, input_data, message
+    capsys, nursery, mock_input, input_data, message
 ):
-    mocker.patch('ws.console.console.input', get_fake_input(input_data))
+    mock_input.send_text(f'{input_data}\nquit\n')
     await nursery.start(serve_websocket, server_handler, 'localhost', 1234, None)
     await main('ws://localhost:1234')
     output = capsys.readouterr().out
@@ -21,9 +21,9 @@ async def test_should_print_unknown_arguments_when_they_are_passed_to_pong_sub_c
     assert 'Bye!' in output
 
 
-async def test_should_print_error_message_when_payload_length_exceeds_125(capsys, mocker, nursery):
+async def test_should_print_error_message_when_payload_length_exceeds_125(capsys, nursery, mock_input):
     payload = 'a' * 126
-    mocker.patch('ws.console.console.input', get_fake_input(f'pong {payload}'))
+    mock_input.send_text(f'pong {payload}\nquit\n')
     await nursery.start(serve_websocket, server_handler, 'localhost', 1234, None)
     await main('ws://localhost:1234')
     output = capsys.readouterr().out
@@ -36,9 +36,9 @@ async def test_should_print_error_message_when_payload_length_exceeds_125(capsys
 @pytest.mark.parametrize(
     ('input_data', 'length'), [('pong', 0), ('pong "hello world"', 11), ("pong 'hello world'", 11)]
 )
-async def test_should_print_success_message_when_pong_is_sent(capsys, mocker, nursery, input_data, length):
+async def test_should_print_success_message_when_pong_is_sent(capsys, nursery, mock_input, input_data, length):
     url = 'ws://localhost:1234'
-    mocker.patch('ws.console.console.input', get_fake_input(input_data))
+    mock_input.send_text(f'{input_data}\nquit\n')
     await nursery.start(serve_websocket, server_handler, 'localhost', 1234, None)
     await main(url)
     output = capsys.readouterr().out
