+++ tests/commands/session/test_quit.py
@@ -4,8 +4,8 @@ from tests.helpers import server_handler
 from ws.commands.session import interact
 
 
-async def test_should_print_documentation_and_exit_program(capsys, mocker, nursery):
-    mocker.patch('ws.console.console.input', lambda prompt: 'quit')
+async def test_should_print_documentation_and_exit_program(capsys, nursery, mock_input):
+    mock_input.send_text('quit\n')
     await nursery.start(serve_websocket, server_handler, 'localhost', 1234, None)
     await interact('ws://localhost:1234')
